refactor(server): split http wiring into modules
